llm_tools/gemini_example.py
from dotenv import load_dotenv
from langchain_google_genai import ChatGoogleGenerativeAI
from langchain_core.messages import HumanMessage

load_dotenv()

llm = ChatGoogleGenerativeAI(model="gemini-2.5-flash")

response = llm.invoke([HumanMessage(content="Hello! What can you do?")])
print(response.content)
